requirements.txt
pillow>=5.0.0
numpy>=1.16.0